requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=1.9.0
transformers>=4.11.0
sentence-transformers>=2.1.0
spacy>=3.1.0
nltk>=3.6.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
openai>=0.27.0
python-dotenv>=0.19.0
tqdm>=4.62.0 